requirements.txt
tensorflow_hub
tensorflow
numpy
streamlit
pillow
torch
transformers == 4.46.0
torchvision